cas-main/requirements.txt
click==8.1.3
cmake==3.22.5
eventlet==0.30.2
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.1.2
Flask-SocketIO==4.3.1
grpcio==1.46.3
gunicorn==20.1.0
numpy==1.22.4
opencv-python 
pandas==1.4.3
pandas-compat==0.1.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
python-dateutil==2.8.2
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.28.0
Pillow
redis
six
Werkzeug==2.0.0
itsdangerous
Jinja2
MarkupSafe
face-recognition==1.3.0
face-recognition-models==0.3.0





